Add wishlist functionality and UI
A new service 'WishlistService' has been added which implements methods to fetch, add, and remove items from the wishlist by interacting with the server. The 'WishlistComponent' now uses this service to load wishlist items and allows to add them to the cart. Additionally, a new UI for the wishlist has been added showcasing the items with their cover image, title, price and an 'Add to Cart' button.